% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{BIA_Mar23_Table3}
\alias{BIA_Mar23_Table3}
\title{Business Indicators Australia, March 2023, Table 3}
\format{
A list containing the downloaded and cleaned series data and metadata
\describe{
\item{ data }{ A tibble containing the original, seasonally adjusted and trend series }
\item{ meta }{ A tibble containing the associated metadata for each series }
}
}
\source{
Australian Bureau of Statistics (https://www.abs.gov.au/)
}
\usage{
BIA_Mar23_Table3
}
\description{
Business Indicators Australia, March 2023, Table 3
}
\keyword{datasets}
