% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{check_series}
\alias{check_series}
\title{Check if a series object is in the expected format}
\usage{
check_series(series)
}
\arguments{
\item{series}{The object to be checked}
}
\description{
\code{check_series} checks if an object conforms to the format as per the list
returned from \code{download_ts} containing series data and metadata.
}
\details{
An error is thrown if the object does not conform to the required format.
Otherwise the function returns nothing.
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/labour/" # ABS website (for labour statistics)
pub <- "employment-and-unemployment/labour-force-australia/" # Publication
ref <- "jul-2023/" # Reference period
cat <- "620200" # Catalogue number
tab <- "1" # Table number
ext <- ".xlsx" # File extension
URL <- paste0(ABS, pub, ref, cat, tab, ext)

# Run function
emp <- download_ts(URL)
check_series(emp)
}

}
