% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{CPI}
\alias{CPI}
\title{Consumer Price Index Australia, June Quarter 2023, Tables 1 and 2}
\format{
A list containing the downloaded and cleaned series data and metadata
\describe{
\item{data}{A tibble containing the original, seasonally adjusted and trend
series}
\item{meta}{A tibble containing the associated metadata for each series}
}
}
\source{
\url{https://www.abs.gov.au/statistics/economy/price-indexes-and-inflation/consumer-price-index-australia/jun-quarter-2023#data-downloads}
}
\usage{
CPI
}
\description{
A dataset containing quarterly consumer price index numbers and percentage
changes collected by the
\href{https://www.abs.gov.au/}{Australian Bureau of Statistics}.
}
\keyword{datasets}
