% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_groups.R
\name{create_ts}
\alias{create_ts}
\title{create_ts Create time series objects from downloaded and cleaned ABS time series data}
\usage{
create_ts(series, name)
}
\arguments{
\item{series}{A list returned from download_ts containing series and their metadata}

\item{name}{A series ID}
}
\value{
A time series object
}
\description{
create_ts Create time series objects from downloaded and cleaned ABS time series data
}
\details{
This function creates the time series objects ready for extraction, seasonal adjustment or trending, or plotting.
}
