% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_ts_comp}
\alias{create_ts_comp}
\title{create_ts_comp Create a list of time series objects for one group of original, seasonally adjusted and trend series corresponding to a single data item description in a downloaded and cleaned table from an ABS publication in addition to seasonally adjusted and trend series from X-13ARIMA-SEATS and JDemetra+}
\usage{
create_ts_comp(series, dataItem)
}
\arguments{
\item{series}{A list returned from download_ts containing series and their metadata}

\item{dataItem}{The data item description for which results are to be returned}
}
\value{
A list for a single data item description containing (as time series objects) the ABS original, seasonally adjusted and trend estimates and associated metadata, along with the seasonally adjusted and trend estimates generated by X-13ARIMA-SEATS and JDemetra+
}
\description{
create_ts_comp Create a list of time series objects for one group of original, seasonally adjusted and trend series corresponding to a single data item description in a downloaded and cleaned table from an ABS publication in addition to seasonally adjusted and trend series from X-13ARIMA-SEATS and JDemetra+
}
\details{
Requires the data output from download_ts. If the ABS seasonally adjusted and/or trend estimates are not published, then they will not exist in the list (rather than assume NULL values, for instance).
}
