% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_groups.R
\name{create_ts_groups}
\alias{create_ts_groups}
\title{create_ts_groups Create matching groups of series IDs}
\usage{
create_ts_groups(series)
}
\arguments{
\item{series}{A list returned from download_ts containing series and their metadata}
}
\value{
A list containing character vectors of matching groups of original, seasonally adjusted and trend series IDs
}
\description{
create_ts_groups Create matching groups of series IDs
}
\details{
This function creates matching groups of series IDs from the metadata ('Index' tab of the downloaded ABS spreadsheet) corresponding to original, seasonally adjusted and trend series (in the 'Data1' tab of the spreadsheet). This is necessary to ensure matching series are plotted together by data item description. Note though that seasonally adjusted and trend series are not necessarily available for all original time series.
}
