% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_ts_table}
\alias{create_ts_table}
\title{create_ts_table Create a list of time series objects for all original, seasonally adjusted and trend series in a downloaded and cleaned table from an ABS publication in addition to seasonally adjusted and trend series from X-13ARIMA-SEATS and JDemetra+}
\usage{
create_ts_table(series)
}
\arguments{
\item{series}{A list returned from download_ts containing series and their metadata}
}
\value{
A list for all data item descriptions containing the ABS original, seasonally adjusted and trend estimates and associated metadata, along with the seasonally adjusted and trend estimates generated by X-13ARIMA-SEATS and JDemetra+
}
\description{
create_ts_table Create a list of time series objects for all original, seasonally adjusted and trend series in a downloaded and cleaned table from an ABS publication in addition to seasonally adjusted and trend series from X-13ARIMA-SEATS and JDemetra+
}
\details{
Requires the data output from download_ts. Null values appear in place of non-existent series in the list returned.
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/economy/" # ABS website (for economic statistics)
pub <- "business-indicators/business-indicators-australia/" # Publication
ref <- "mar-2023/" # Reference period
cat <- "567600" # Catalogue number
tab <- "3.xlsx" # Table number
URL <- paste0(ABS, pub, ref, cat, tab)

# Download and plot data
sales <- download_ts(URL)
}
}
