% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsdf_comp}
\alias{create_tsdf_comp}
\title{Generate a long data frame of original, seasonally adjusted and trend
estimates for a single data item description}
\usage{
create_tsdf_comp(series, dataItem)
}
\arguments{
\item{series}{The list returned from \code{download_ts} containing series data and
metadata}

\item{dataItem}{The data item description for which results are to be returned}
}
\value{
The series metadata, and long data frame of original, seasonally adjusted
and trend series, for a single data item description
}
\description{
\code{create_tsdf_comp} returns a list containing the following for a single data
item description:
\itemize{
\item the series metadata
\item a long data frame containing the original, seasonally adjusted and trend
estimates from the ABS, X-13ARIMA-SEATS, and JDemetra+ (where available)
}
}
\details{
Requires the data output from \code{download_ts}. If the ABS seasonally adjusted
and/or trend estimates are not published, then they will not exist in the
long data frame (rather than assume NULL values, for instance).

This function is useful for converting the data into a format ready to be
plotted. (For example, plotting with ggplot2 requires data frames, not time
series objects.)
}
\examples{
# Load data
data("inventories")

# Create long data frame for a given data item description
item <- "Inventories ;  Total (State) ;  Mining ;  Current Price ;  TOTAL (SCP_SCOPE) ;"
inventories_df_mining <- create_tsdf_comp(inventories, item)

}
