% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsdf_table}
\alias{create_tsdf_table}
\title{Create a list of long data frames for all original, seasonally adjusted and trend series in a downloaded and cleaned table from an ABS publication in addition to seasonally adjusted and trend series from X-13ARIMA-SEATS and JDemetra+}
\usage{
create_tsdf_table(series)
}
\arguments{
\item{series}{A list returned from download_ts containing series and their metadata}
}
\value{
A list for all data item descriptions containing (as long data frames) the ABS original, seasonally adjusted and trend estimates and associated metadata, along with the seasonally adjusted and trend estimates generated by X-13ARIMA-SEATS and JDemetra+
}
\description{
Create a list of long data frames for all original, seasonally adjusted and trend series in a downloaded and cleaned table from an ABS publication in addition to seasonally adjusted and trend series from X-13ARIMA-SEATS and JDemetra+
}
\details{
Requires the data output from download_ts. Null values appear in place of non-existent series in the list returned.
}
