% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsplot_comp}
\alias{create_tsplot_comp}
\title{Generate a ggplot for seasonally adjusted or trend estimates for a single
data item description}
\usage{
create_tsplot_comp(series, dataItem, type)
}
\arguments{
\item{series}{list returned from \code{download_ts} containing series data and
metadata}

\item{dataItem}{data item description for which results are to be returned}

\item{type}{"S" for seasonal adjustment, or "T" for trending}
}
\value{
ggplot object for the seasonally adjusted or trend series for a
single data item description
}
\description{
\code{create_tsplot_comp} returns a ggplot object for either the seasonally
adjusted or trend series corresponding to a single data item description in
a downloaded and cleaned table from an ABS publication (as returned by
\code{download_ts}).
}
\details{
Requires the data output from \code{download_ts}. If the ABS seasonally adjusted
or trend estimates are not published, then a ggplot is still returned
containing only the seasonally adjusted or trend estimates generated by
X-13ARIMA-SEATS and JDemetra+.
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/labour/" # ABS website (for labour statistics)
pub <- "employment-and-unemployment/labour-force-australia/" # Publication
ref <- "jul-2023/" # Reference period
cat <- "620200" # Catalogue number
tab <- "1" # Table number
ext <- ".xlsx" # File extension
URL <- paste0(ABS, pub, ref, cat, tab, ext)

# Run function
create_tsplot_comp(emp, "Employed total ;  Persons ;", "S")

# Plot still produced even when ABS seasonally adjusted and/or trend non-existent:
create_tsplot_comp(emp, "Civilian population aged 15 years and over ;  Persons ;", "S")
}

}
