% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsplot_table}
\alias{create_tsplot_table}
\title{Generate ggplots for seasonally adjusted and trend estimates for all data
item descriptions}
\usage{
create_tsplot_table(series)
}
\arguments{
\item{series}{list returned from \code{download_ts} containing series data and
metadata}
}
\value{
ggplot objects for the seasonally adjusted or trend series for all
data item descriptions
}
\description{
\code{create_tsplot_table} returns a list of ggplot objects for the seasonally
adjusted and trend series corresponding to all data item descriptions in a
downloaded and cleaned table from an ABS publication (as returned by
\code{download_ts}).
}
\details{
Requires the data output from \code{download_ts}. If the ABS seasonally adjusted
and/or trend estimates are not published, then a ggplot is still returned
containing only the seasonally adjusted or trend estimates generated by
X-13ARIMA-SEATS and JDemetra+.
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/labour/" # ABS website (for labour statistics)
pub <- "employment-and-unemployment/labour-force-australia/" # Publication
ref <- "jul-2023/" # Reference period
cat <- "620200" # Catalogue number
tab <- "1" # Table number
ext <- ".xlsx" # File extension
URL <- paste0(ABS, pub, ref, cat, tab, ext)

# Run function
emp_plots <- create_tsplot_table(emp)
}

}
