% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{Australian National Accounts: National Income, Expenditure and Product, June
Quarter 2023, Table 1}
\format{
A list containing the downloaded and cleaned series data and metadata
\describe{
\item{data}{A tibble containing the original, seasonally adjusted and trend
series}
\item{meta}{A tibble containing the associated metadata for each series}
}
}
\source{
\url{https://www.abs.gov.au/statistics/economy/national-accounts/australian-national-accounts-national-income-expenditure-and-product/jun-2023#data-downloads}
}
\usage{
GDP
}
\description{
A dataset containing quarterly key national accounts aggregates collected by
the \href{https://www.abs.gov.au/}{Australian Bureau of Statistics}.
}
\keyword{datasets}
