% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{get_dataItems}
\alias{get_dataItems}
\title{Get all data item descriptions from series metadata}
\usage{
get_dataItems(series)
}
\arguments{
\item{series}{The list returned from \code{download_ts} containing series data and
metadata}
}
\value{
All data item descriptions in the metadata of a series
}
\description{
\code{get_dataItems} returns all unique data item descriptions in the series
metadata for a downloaded and cleaned table from an ABS publication (as
returned by \code{download_ts}).
}
\details{
The data item descriptions are returned in an unsorted character vector, thus
preserving the original ordering in the series metadata ('Index' sheet of the
downloaded spreadsheet) for series that remain after data cleaning.
}
\examples{
# Load data
data("inventories")

# Extract all data item descriptions
get_dataItems(inventories)

}
