% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{get_seriesIDs}
\alias{get_seriesIDs}
\title{Get all series IDs from series metadata}
\usage{
get_seriesIDs(series)
}
\arguments{
\item{series}{list returned from \code{download_ts} containing series data and
metadata}
}
\value{
all series IDs in the metadata of a series
}
\description{
\code{get_seriesIDs} returns all series IDs in the series metadata for a
downloaded and cleaned table from an ABS publication (as returned by
\code{download_ts}).
}
\details{
The series IDs are returned in an unsorted character vector, thus
preserving the original ordering in the series metadata ('Index' sheet of the
downloaded spreadsheet) for series that remain after data cleaning.
}
