% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{LFA_Jul23_Table1}
\alias{LFA_Jul23_Table1}
\title{Labour Force Australia, July 2023, Table 1}
\format{
A list containing the downloaded and cleaned series data and metadata
\describe{
\item{data}{A tibble containing the original, seasonally adjusted and trend
series}
\item{meta}{A tibble containing the associated metadata for each series}
}
}
\source{
\url{https://www.abs.gov.au/statistics/labour/employment-and-unemployment/labour-force-australia/jul-2023#data-downloadss}
}
\usage{
LFA_Jul23_Table1
}
\description{
A dataset containing monthly estimates of labour force status by sex
collected by the \href{https://www.abs.gov.au/}{Australian Bureau of Statistics}.
}
\keyword{datasets}
