% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{create_ts Create time series objects from downloaded and cleaned ABS time series data}
\usage{
plot_ts(series, type = "S")
}
\arguments{
\item{series}{A list returned from download_ts containing series and their metadata}

\item{type}{Optional: "S" for seasonally adjusted plots, or "T" for trend plots. Defaults to "S"}
}
\value{
A list of ggplot objects
}
\description{
create_ts Create time series objects from downloaded and cleaned ABS time series data
}
\details{
Requires the data output from download_ts, and calls the time series generation functions (create_groups and create_ts) during execution. Where an original series is not accompanied by seasonally adjusted or trend series, it will not be plotted as there is nothing to compare.
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/economy/" # ABS website (for economic statistics)
pub <- "business-indicators/business-indicators-australia/" # Publication
ref <- "mar-2023/" # Reference period
cat <- "567600" # Catalogue number
tab <- "3.xlsx" # Table number
URL <- paste0(ABS, pub, ref, cat, tab)

# Download and plot data
sales <- download_ts(URL)
sales_plots <- plot_ts(sales, "T")
for (p in sales_plots) print(p)
}
}
