% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{retail}
\alias{retail}
\title{Retail Trade Australia, July 2023, Table 1}
\format{
A list containing the downloaded and cleaned series data and metadata
\describe{
\item{data}{A tibble containing the original, seasonally adjusted and trend
series}
\item{meta}{A tibble containing the associated metadata for each series}
}
}
\source{
\url{https://www.abs.gov.au/statistics/industry/retail-and-wholesale-trade/retail-trade-australia/jul-2023#data-downloads}
}
\usage{
retail
}
\description{
A dataset containing monthly estimates of retail turnover by industry group
collected by the \href{https://www.abs.gov.au/}{Australian Bureau of Statistics}.
}
\keyword{datasets}
