% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{run_RJD}
\alias{run_RJD}
\title{run_RJD Run JDemetra+ to generate seasonally adjusted and trend estimates for a time series object}
\usage{
run_RJD(ts)
}
\arguments{
\item{ts}{A time series object containing original estimates}
}
\value{
A list of time series object containing seasonally adjusted and trend estimates generated by JDemetra+
}
\description{
run_RJD Run JDemetra+ to generate seasonally adjusted and trend estimates for a time series object
}
\details{
Seasonal adjustment and trending are performed using X-13 rather than SEATS
}
