% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{run_X13}
\alias{run_X13}
\title{run_X13 Run X-13ARIMA-SEATS to generate seasonally adjusted or trend estimates for a time series object}
\usage{
run_X13(ts, type)
}
\arguments{
\item{ts}{A time series object containing original estimates}

\item{type}{Optional: "S" for seasonal adjustment, or "T" for trending}
}
\value{
A time series object containing seasonally adjusted or trend estimates generated by X-13ARIMA-SEATS
}
\description{
run_X13 Run X-13ARIMA-SEATS to generate seasonally adjusted or trend estimates for a time series object
}
\details{
Seasonal adjustment and trending are performed using X-13 rather than SEATS
}
