% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{sales}
\alias{sales}
\title{Business Indicators Australia, June Quarter 2023, Table 6}
\format{
A list containing the downloaded and cleaned series data and metadata
\describe{
\item{data}{A tibble containing the original, seasonally adjusted and trend
series}
\item{meta}{A tibble containing the associated metadata for each series}
}
}
\source{
\url{https://www.abs.gov.au/statistics/economy/business-indicators/business-indicators-australia/jun-2023#data-downloads}
}
\usage{
sales
}
\description{
A dataset containing quarterly estimates of private sector income from sales
of goods and services in current prices collected by the
\href{https://www.abs.gov.au/}{Australian Bureau of Statistics}.
}
\keyword{datasets}
