% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{BIA_Mar23_Table3}
\alias{BIA_Mar23_Table3}
\title{Business Indicators Australia, March 2023, Table 3}
\format{
A list containing the downloaded and cleaned series data and metadata
\describe{
\item{data}{A tibble containing the original, seasonally adjusted and trend
series}
\item{meta}{A tibble containing the associated metadata for each series}
}
}
\source{
\url{https://www.abs.gov.au/statistics/economy/business-indicators/business-indicators-australia/mar-2023#data-downloads}
}
\usage{
BIA_Mar23_Table3
}
\description{
A dataset containing quarterly estimates of private sector inventories in
current prices collected by the
\href{https://www.abs.gov.au/}{Australian Bureau of Statistics}.
}
\keyword{datasets}
