% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{check_series}
\alias{check_series}
\title{Check if a series object is in the expected format}
\usage{
check_series(series)
}
\arguments{
\item{series}{The object to be checked}
}
\description{
\code{check_series} checks if an object conforms to the format as per the list
returned from \code{download_ts} containing series data and metadata.
}
\details{
An error is thrown if the object does not conform to the required format.
Otherwise the function returns nothing. Only limited error checking is
undertaken, such as checking for appropriate variable names and types.
Logical checks for aspects such as consistency within or between the series
data and metadata are not performed.
}
\examples{
# Load data
data("inventories")

# Check the series object for validity (nothing happens if valid)
check_series(inventories)

}
