% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_ts}
\alias{create_ts}
\title{Create a time series object for a single series}
\usage{
create_ts(series, seriesID)
}
\arguments{
\item{series}{The list returned from \code{download_ts} containing series data and
metadata}

\item{seriesID}{The series ID of the series for which results are to be
returned}
}
\value{
A time series object for a single series
}
\description{
\code{create_ts} returns a time series object for a single series in a downloaded
and cleaned table from an ABS publication (as returned by \code{download_ts}).
The series is identified by its series ID contained in the series metadata.
}
\details{
To maintain consistency with other structures and functions used in this
package, a NULL object (rather than an error) is returned if the series ID is
missing.
}
\examples{
# Load data
data("inventories")

# Create a time series object for a given series ID
inventories_mining_orig <- create_ts(inventories, "A3531252X")

}
