% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_ts}
\alias{create_ts}
\title{create_ts Create a time series object for a single series in a downloaded and cleaned table from an ABS publication}
\usage{
create_ts(series, seriesID)
}
\arguments{
\item{series}{A list returned from download_ts containing series and their metadata}

\item{seriesID}{A series ID}
}
\value{
A time series object for a single series
}
\description{
create_ts Create a time series object for a single series in a downloaded and cleaned table from an ABS publication
}
