% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_ts}
\alias{create_ts}
\title{Create a time series object for a single series}
\usage{
create_ts(series, seriesID)
}
\arguments{
\item{series}{list returned from \code{download_ts} containing series data and
metadata}

\item{seriesID}{series ID}
}
\value{
time series object for a single series
}
\description{
\code{create_ts} returns a time series object for a single series in a downloaded
and cleaned table from an ABS publication (as returned by \code{download_ts}).
The series is identified by its series ID contained in the series metadata.
}
\details{
To maintain consistency with other structures and functions used in this
package, a NULL object (rather than an error) is returned if the series ID is
missing.
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/labour/" # ABS website (for labour statistics)
pub <- "employment-and-unemployment/labour-force-australia/" # Publication
ref <- "jul-2023/" # Reference period
cat <- "620200" # Catalogue number
tab <- "1" # Table number
ext <- ".xlsx" # File extension
URL <- paste0(ABS, pub, ref, cat, tab, ext)

# Run function
emp_total <- create_ts(emp, "A84423127L")
}

}
