% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_ts_table}
\alias{create_ts_table}
\title{Generate original, seasonally adjusted and trend estimates for all data
item descriptions}
\usage{
create_ts_table(series)
}
\arguments{
\item{series}{The list returned from \code{download_ts} containing series data and
metadata}
}
\value{
The series metadata, and original, seasonally adjusted and trend series,
for all data item descriptions
}
\description{
\code{create_ts_table} returns a list of lists, each sub-list of which corresponds
to a single data item description, and contains the following for each one:
\itemize{
\item the series metadata
\item the original ABS series
\item the seasonally adjusted and trend ABS series (where available)
\item the seasonally adjusted and trend series generated by X-13ARIMA-SEATS and
JDemetra+
}
}
\details{
Requires the data output from \code{download_ts}. If the ABS seasonally adjusted
and/or trend estimates are not published, then they will not exist in the
lists (rather than assume NULL values, for instance).
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/labour/" # ABS website (for labour statistics)
pub <- "employment-and-unemployment/labour-force-australia/" # Publication
ref <- "jul-2023/" # Reference period
cat <- "620200" # Catalogue number
tab <- "1" # Table number
ext <- ".xlsx" # File extension
URL <- paste0(ABS, pub, ref, cat, tab, ext)

# Run function
emp_all <- create_ts_table(emp)
}

}
