% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsdf_comp}
\alias{create_tsdf_comp}
\title{create_tsdf_comp Create a long data frame for one group of original, seasonally adjusted and trend series corresponding to a single data item description in a downloaded and cleaned table from an ABS publication in addition to seasonally adjusted and trend series from X-13ARIMA-SEATS and JDemetra+}
\usage{
create_tsdf_comp(series, dataItem)
}
\arguments{
\item{series}{A list returned from download_ts containing series and their metadata}

\item{dataItem}{The data item description for which results are to be returned}
}
\value{
A list for a single data item description containing (as a long data frame) the ABS original, seasonally adjusted and trend estimates and associated metadata, along with the seasonally adjusted and trend estimates generated by X-13ARIMA-SEATS and JDemetra+
}
\description{
create_tsdf_comp Create a long data frame for one group of original, seasonally adjusted and trend series corresponding to a single data item description in a downloaded and cleaned table from an ABS publication in addition to seasonally adjusted and trend series from X-13ARIMA-SEATS and JDemetra+
}
\details{
This function is useful for converting the data into a format ready to be plotted. (For example, plotting with ggplot2 requires data frames, not time series objects.) Requires the data output from download_ts. If the ABS seasonally adjusted and/or trend estimates are not published, then they will not exist in the long data frame (rather than assume NULL values, for instance).
}
