% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsdf_table}
\alias{create_tsdf_table}
\title{Generate a long data frame of original, seasonally adjusted and trend
estimates for all data item descriptions}
\usage{
create_tsdf_table(series)
}
\arguments{
\item{series}{The list returned from \code{download_ts} containing series data and
metadata}
}
\value{
The series metadata, and long data frame of original, seasonally adjusted
and trend series, for all data item descriptions
}
\description{
\code{create_tsdf_table} returns a list of lists, each sub-list of which
corresponds to a single data item description, and contains the following for
each one:
\itemize{
\item the series metadata
\item a long data frame containing the original, seasonally adjusted and trend
estimates from the ABS, X-13ARIMA-SEATS, and JDemetra+ (where available)
}
}
\details{
Requires the data output from \code{download_ts}. If the ABS seasonally adjusted
and/or trend estimates are not published, then they will not exist in the
long data frames (rather than assume NULL values, for instance).
}
\examples{
# Load data
data("inventories")

# Create long data frame for all data item descriptions
inventories_df_all <- create_tsdf_table(inventories)

}
