% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsdf_table}
\alias{create_tsdf_table}
\title{Generate a long data frame of original, seasonally adjusted and trend
estimates for all data item descriptions}
\usage{
create_tsdf_table(series)
}
\arguments{
\item{series}{list returned from \code{download_ts} containing series data and
metadata}
}
\value{
series metadata, and long data frame of original, seasonally adjusted
and trend series, for all data item descriptions
}
\description{
\code{create_tsdf_table} returns a list of lists, each sub-list of which
corresponds to a single data item description, and contains the following for
each one:
\itemize{
\item the series metadata
\item a long data frame containing the original, seasonally adjusted and trend
estimates from the ABS, X-13ARIMA-SEATS, and JDemetra+ (where available)
}
}
\details{
Requires the data output from \code{download_ts}. NULL values appear in place of
non-existent series in the list returned.
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/labour/" # ABS website (for economic statistics)
pub <- "employment-and-unemployment/labour-force-australia/" # Publication
ref <- "jul-2023/" # Reference period
cat <- "620200" # Catalogue number
tab <- "1.xlsx" # Table number
URL <- paste0(ABS, pub, ref, cat, tab)

# Run function
emp_all <- create_tsdf_table(emp)
}

}
