% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsplot_comp}
\alias{create_tsplot_comp}
\title{Generate ggplots for seasonally adjusted and trend estimates for a single
data item description}
\usage{
create_tsplot_comp(series, dataItem)
}
\arguments{
\item{series}{The list returned from \code{download_ts} containing series data and
metadata}

\item{dataItem}{The data item description for which results are to be returned}
}
\value{
The ggplot objects for the seasonally adjusted and trend series for a
single data item description
}
\description{
\code{create_tsplot_comp} returns a list of ggplot objects for the seasonally
adjusted and trend series corresponding to a single data item description in
a downloaded and cleaned table from an ABS publication (as returned by
\code{download_ts}). For each series type (seasonally adjusted and trend), the
following series will be plotted on a single graph:
\itemize{
\item the ABS seasonally adjusted/trend series (if published)
\item the seasonally adjusted/trend series generated by X-13ARIMA-SEATS (if
generated without error)
\item the seasonally adjusted/trend series generated by JDemetra+ (if generated
without error)
}
}
\details{
Requires the data output from \code{download_ts}. If the ABS seasonally adjusted
or trend estimates are not published, then a ggplot is still returned
containing only the seasonally adjusted or trend estimates generated by
X-13ARIMA-SEATS and JDemetra+. If the latter fail to be generated, the ggplot
will not contain them either. (A NULL object will be returned if the plot is
empty; that is, if none of these series are published or couldn't be
generated.)

The time series plotted on each graph might not match spans or have values
missing for the following reasons:
\itemize{
\item suppressed or missing observations in the published ABS series
\item series generated by X-13ARIMA-SEATS and JDemetra+ being restricted to only
recent observations due to the ABS original time series being too long (see
the \code{download_ts} function)
}
}
\examples{
# Load data
data("inventories")

# Create plots for a given data item description
item <- "Inventories ;  Total (State) ;  Mining ;  Current Price ;  TOTAL (SCP_SCOPE) ;"
inventories_mining_plots <- create_tsplot_comp(inventories, item)

}
