% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsplot_comp}
\alias{create_tsplot_comp}
\title{Create a ggplot object for either the seasonally adjusted or trend series corresponding to a single data item description in a downloaded and cleaned table from an ABS publication}
\usage{
create_tsplot_comp(series, dataItem, type)
}
\arguments{
\item{series}{A list returned from download_ts containing series and their metadata}

\item{dataItem}{The data item description for which results are to be returned}

\item{type}{Optional: "S" for seasonal adjustment, or "T" for trending}
}
\value{
A ggplot object for either the seasonally adjusted or trend series
}
\description{
Create a ggplot object for either the seasonally adjusted or trend series corresponding to a single data item description in a downloaded and cleaned table from an ABS publication
}
\details{
Requires the data output from download_ts. If the ABS seasonally adjusted and/or trend estimates are not published, then a NULL object will be returned.
}
