% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsplot_table}
\alias{create_tsplot_table}
\title{Generate ggplots for seasonally adjusted and trend estimates for all data
item descriptions}
\usage{
create_tsplot_table(series)
}
\arguments{
\item{series}{The list returned from \code{download_ts} containing series data and
metadata}
}
\value{
The ggplot objects for the seasonally adjusted or trend series for all
data item descriptions
}
\description{
\code{create_tsplot_table} returns a list of ggplot objects for the seasonally
adjusted and trend series corresponding to all data item descriptions in a
downloaded and cleaned table from an ABS publication (as returned by
\code{download_ts}). For each series type (seasonally adjusted and trend) for each
data item description, the following series will be plotted on a single
graph:
\itemize{
\item the ABS seasonally adjusted/trend series (if published)
\item the seasonally adjusted/trend series generated by X-13ARIMA-SEATS (if
generated without error)
\item the seasonally adjusted/trend series generated by JDemetra+ (if generated
without error)
}
}
\details{
Requires the data output from \code{download_ts}. If the ABS seasonally adjusted
or trend estimates are not published for a given data item description, then
a ggplot is still returned containing only the seasonally adjusted or trend
estimates generated by X-13ARIMA-SEATS and JDemetra+. If the latter fail to
be generated, the ggplot will not contain them either. (A NULL object will be
returned instead if the plot is empty; that is, if none of these series are
published or couldn't be generated.)

The time series plotted on each graph for a given data item description might
not match spans or have values missing for the following reasons:
\itemize{
\item suppressed or missing observations in the published ABS series
\item series generated by X-13ARIMA-SEATS and JDemetra+ being restricted to only
recent observations due to the ABS original time series being too long (see
the \code{download_ts} function)
}
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/labour/" # ABS website (for labour statistics)
pub <- "employment-and-unemployment/labour-force-australia/" # Publication
ref <- "jul-2023/" # Reference period
cat <- "620200" # Catalogue number
tab <- "1" # Table number
ext <- ".xlsx" # File extension
URL <- paste0(ABS, pub, ref, cat, tab, ext)

# Run function
emp_plots <- create_tsplot_table(emp)
}

}
