% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{create_tsplot_table}
\alias{create_tsplot_table}
\title{Create a list of ggplot objects for the seasonally adjusted and trend series for all data item descriptions in a downloaded and cleaned table from an ABS publication}
\usage{
create_tsplot_table(series)
}
\arguments{
\item{series}{A list returned from download_ts containing series and their metadata}
}
\value{
A list of ggplot objects for the seasonally adjusted and trend series for all data item descriptions
}
\description{
Create a list of ggplot objects for the seasonally adjusted and trend series for all data item descriptions in a downloaded and cleaned table from an ABS publication
}
\details{
Requires the data output from download_ts. The ggplot objects containing the seasonally adjusted and trend estimates generated by X-13ARIMA-SEATS and JDemetra+ are still created for ABS data where only the original estimates are available.
}
