% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{get_dataItems}
\alias{get_dataItems}
\title{Get all data item descriptions from series metadata}
\usage{
get_dataItems(series)
}
\arguments{
\item{series}{The list returned from \code{download_ts} containing series data and
metadata}
}
\value{
All data item descriptions in the metadata of a series
}
\description{
\code{get_dataItems} returns all unique data item descriptions in the series
metadata for a downloaded and cleaned table from an ABS publication (as
returned by \code{download_ts}).
}
\details{
The data item descriptions are returned in an unsorted character vector, thus
preserving the original ordering in the series metadata ('Index' sheet of the
downloaded spreadsheet) for series that remain after data cleaning.
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/labour/" # ABS website (for labour statistics)
pub <- "employment-and-unemployment/labour-force-australia/" # Publication
ref <- "jul-2023/" # Reference period
cat <- "620200" # Catalogue number
tab <- "1" # Table number
ext <- ".xlsx" # File extension
URL <- paste0(ABS, pub, ref, cat, tab, ext)

# Run function
emp <- download_ts(URL)
get_dataItems(emp)
}

}
