% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{get_seriesIDs}
\alias{get_seriesIDs}
\title{Get all series IDs from series metadata}
\usage{
get_seriesIDs(series)
}
\arguments{
\item{series}{The list returned from \code{download_ts} containing series data and
metadata}
}
\value{
All series IDs in the metadata of a series
}
\description{
\code{get_seriesIDs} returns all series IDs in the series metadata for a
downloaded and cleaned table from an ABS publication (as returned by
\code{download_ts}).
}
\details{
The series IDs are returned in an unsorted character vector, thus
preserving the original ordering in the series metadata ('Index' sheet of the
downloaded spreadsheet) for series that remain after data cleaning.
}
\examples{
\dontrun{
# Construct URL to download ABS data
ABS <- "https://www.abs.gov.au/statistics/labour/" # ABS website (for labour statistics)
pub <- "employment-and-unemployment/labour-force-australia/" # Publication
ref <- "jul-2023/" # Reference period
cat <- "620200" # Catalogue number
tab <- "1" # Table number
ext <- ".xlsx" # File extension
URL <- paste0(ABS, pub, ref, cat, tab, ext)

# Run function
emp <- download_ts(URL)
get_seriesIDs(emp)
}

}
