% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{run_RJD}
\alias{run_RJD}
\title{Generate seasonally adjusted and trend estimates via JDemetra+}
\usage{
run_RJD(ts)
}
\arguments{
\item{ts}{A time series object}
}
\value{
The seasonally adjusted and trend estimates generated by JDemetra+
}
\description{
\code{run_RJD} executes the Eurostat's JDemetra+ program to generate the
seasonally adjusted and trend estimates for a time series object (containing
original estimates), and returns these in a list.
}
\details{
Seasonal adjustment and trending are performed using the \code{RJDemetra}
interface to JDemetra+ (version 2). Execution is done using X-13 rather than
SEATS; otherwise all default values are used in the execution call.
}
\examples{
# Load data
data("inventories")

# Create original series
inventories_mining_orig <- create_ts(inventories, "A3531252X")

# Generate seasonally adjusted and trend series
inventories_mining_seas_tren <- run_RJD(inventories_mining_orig)

}
