% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ts.R
\name{run_X13}
\alias{run_X13}
\title{Generate seasonally adjusted and trend estimates via X-13ARIMA-SEATS}
\usage{
run_X13(ts)
}
\arguments{
\item{ts}{time series object}
}
\value{
seasonally adjusted and trend estimates generated by X-13ARIMA-SEATS
}
\description{
\code{run_X13} executes the United States Census Bureau's X-13ARIMA-SEATS program
to generate the seasonally adjusted and trend estimates for a time series
object (containing original estimates), and returns these in a list.
}
\details{
Seasonal adjustment and trending are performed using the \code{seasonal} interface
to X-13ARIMA-SEATS. Execution is done using X-13 rather than SEATS; otherwise
all default values are used in the execution call.
}
